\name{constants}
\alias{constants}
\docType{data}
\title{ Constants }
\description{
  Physical constants
}
\usage{data(Constants)}
\format{
  The format is:
List of 8

}
\details{
}
\source{
  NIST
}
\references{
   wikipedia
}
\examples{
data(constants)
constants
comment(constants$cel)

}
\keyword{datasets}
