% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{lookup}
\alias{lookup}
\title{Lookup for Fundamental Physical Constants}
\usage{
lookup(pattern, cols = c("symbol", "quantity", "type"), ...)
}
\arguments{
\item{pattern}{character string containing a regular expression to be matched
(see \code{\link{grep}}).}

\item{cols}{columns to perform pattern matching (see \code{\link{codata}}).}

\item{...}{additional arguments for \code{\link{grep}}.}
}
\description{
A simple wrapper around \code{\link{grep}} for exploring the CODATA dataset.
}
\examples{
lookup("planck", ignore.case=TRUE)

}
\seealso{
\code{\link{codata}}, \code{\link{syms}}.
}
