% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{syms}
\alias{syms}
\alias{syms_with_errors}
\alias{syms_with_units}
\title{Lists containing all symbols.}
\format{An object of class \code{list} of length 0.}
\usage{
syms

syms_with_errors

syms_with_units
}
\description{
Lazy loaded when used
}
\keyword{datasets}
