% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{codata}
\alias{codata}
\title{CODATA Recommended Values of the Fundamental Physical Constants: 2014}
\format{\code{codata} is a data frame with ... cases (rows) and 6 variables (columns)
named \code{quantity}, \code{symbol}, \code{value}, \code{unit}, \code{rel_uncertainty},
and \code{type}.}
\source{
Mohr, P. J., Newell, D. B. and Taylor, B. N. (2016). CODATA recommended
values of the fundamental physical constants: 2014. \emph{Rev. Mod. Phys.},
88, 035009. \doi{10.1103/RevModPhys.88.035009}.

Mohr, P. J., Newell, D. B. and Taylor, B. N. (2016). CODATA recommended values
of the fundamental physical constants: 2014. \emph{J. Phys. Chem. Ref. Data},
45, 043102. \doi{10.1063/1.4954402}.
}
\usage{
codata
}
\description{
The Committee on Data for Science and Technology (CODATA) is an interdisciplinary
committee of the International Council for Science. The Task Group on Fundamental
Constants periodically provides the internationally accepted set of values of
the fundamental physical constants. This dataset contains the "2014 CODATA"
version, published on 25 June 2015.
}
\seealso{
\code{\link{syms}}, \code{\link{lookup}}.
}
\keyword{datasets}
