% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{package}
\name{constants-package}
\alias{constants-package}
\title{\pkg{constants}: Reference on Constants, Units and Uncertainty}
\description{
This package provides the 2018 version of the CODATA internationally recommended
values of the fundamental physical constants for their use within the \R language.
}
\references{
Eite Tiesinga, Peter J. Mohr, David B. Newell, and Barry N. Taylor (2020).
The 2018 CODATA Recommended Values of the Fundamental Physical Constants
(Web Version 8.1). Database developed by J. Baker, M. Douma, and S. Kotochigova.
Available at https://physics.nist.gov/cuu/Constants/,
National Institute of Standards and Technology, Gaithersburg, MD 20899.
}
\seealso{
\code{\link{codata}}, \code{\link{syms}}, \code{\link{lookup}}.
}
\author{
Iñaki Ucar
}
