% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{package}
\name{constants}
\alias{constants}
\alias{constants-package}
\title{\pkg{constants}: Reference on Constants, Units and Uncertainty}
\description{
CODATA internationally recommended 2014 values of the fundamental physical constants.
}
\author{
Iñaki Ucar
}
