% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{package}
\name{constants}
\alias{constants}
\alias{constants-package}
\title{\pkg{constants}: Reference on Constants, Units and Uncertainty}
\description{
This package provides the 2014 version of the CODATA internationally recommended
values of the fundamental physical constants for their use within the \R language.
}
\references{
Mohr, P. J., Newell, D. B. and Taylor, B. N. (2016). CODATA recommended
values of the fundamental physical constants: 2014. \emph{Rev. Mod. Phys.},
88, 035009. \doi{10.1103/RevModPhys.88.035009}.

Mohr, P. J., Newell, D. B. and Taylor, B. N. (2016). CODATA recommended values
of the fundamental physical constants: 2014. \emph{J. Phys. Chem. Ref. Data},
45, 043102. \doi{10.1063/1.4954402}.
}
\seealso{
\code{\link{codata}}, \code{\link{syms}}, \code{\link{lookup}}.
}
\author{
Iñaki Ucar
}
